authorization on comment update and delete

+++ src/components/comment.jsx
@@ -8,8 +8,10 @@ import axios from 'axios';
 
 const Comment = ({ post, setPost, comment }) => {
     const [openEdit, setOpenEdit] = useState(false)
+    const { user, isLoggedIn, logOut } = useContext(AuthContext);
     const [openEditInput, setOpenEditInput] = useState(false)
     const [newComment, setNewComment] = useState('')
+
     const submitEditedComment = (e) => {
         e.preventDefault()
         axios.put(`http://localhost:3000/comments/update-comment/${comment._id}`, {
@@ -29,16 +31,19 @@ const Comment = ({ post, setPost, comment }) => {
             })
     }
     const deleteComment = (e) => {
-        e.preventDefault()
-        axios.delete(`http://localhost:3000/comments/delete/${comment._id}`)
-            .then(res => {
-                setPost(res.data)
-                setOpenEdit(false)
-            })
-            .catch(err => {
-                console.log(err)
-            })
-
+        if (comment.owner._id !== user._id) {
+            console.log('not the owner')
+        } else {
+            e.preventDefault()
+            axios.delete(`http://localhost:3000/comments/delete/${comment._id}`)
+                .then(res => {
+                    setPost(res.data)
+                    setOpenEdit(false)
+                })
+                .catch(err => {
+                    console.log(err)
+                })
+        }
     }
     const openEditBox = () => {
         if (!openEdit) {
@@ -66,8 +71,8 @@ const Comment = ({ post, setPost, comment }) => {
                 </div>
                 {openEditInput ? <>
                     <form onSubmit={submitEditedComment}>
-                        <input style={{borderRadius:'5px', border:'1px solid'}} onClick={()=>{setOpenEdit(false)}} type="text" value={newComment} onChange={(e) => { setNewComment(e.target.value) }} />
-                        <button style={{padding:'5px', border:'none', borderRadius:'5px'}} type='submit'>submit</button>
+                        <input style={{ borderRadius: '5px', border: '1px solid' }} onClick={() => { setOpenEdit(false) }} type="text" value={newComment} onChange={(e) => { setNewComment(e.target.value) }} />
+                        <button style={{ padding: '5px', border: 'none', borderRadius: '5px' }} type='submit'>submit</button>
                     </form>
                 </> :
                     <p style={{ fontSize: '15px', fontWeight: 'lighter', margin: '2px 5px 5px' }}>{comment.comment}</p>
@@ -79,7 +84,13 @@ const Comment = ({ post, setPost, comment }) => {
                 <FavoriteBorderIcon />
                 <MoreHorizIcon onClick={openEditBox} />
                 {openEdit && <div className='edit-comment-buttons' style={{ display: 'flex', padding: '10px', flexDirection: 'column', margin: '8px' }}>
-                    <button onClick={() => { setOpenEditInput(true) }}><p style={{ textAlign: 'left', margin: '10px' }}>Edit</p></button>
+                    <button onClick={() => {
+                        if (comment.owner._id !== user._id) {
+                            console.log('not the owner')
+                        } else {
+                            setOpenEditInput(true)
+                        }
+                    }}><p style={{ textAlign: 'left', margin: '10px' }}>Edit</p></button>
                     <button onClick={deleteComment}><p style={{ textAlign: 'left', margin: '10px' }}>Delete</p></button>
                 </div>}
             </div>
